+++ common/src/main/java/dev/terminalmc/chatnotify/gui/widget/list/option/NotifOptionList.java
@@ -110,7 +110,6 @@ public class NotifOptionList extends OptionList {
     @Override
     public void onClose() {
         notif.editing = false;
-        notif.autoDisable();
     }
 
     private void openTriggerConfig(Trigger trigger) {